remove test utility module
