+++ src/nro45data/psw/ms2/_casa/__init__.py
@@ -8,6 +8,7 @@ if _is_casa6_available:
     from ._casa6 import convert_str_angle_to_rad
     from ._casa6 import put_table_keyword
     from ._casa6 import datestr2mjd
+    from ._casa6 import mjd2datetime
 elif _is_casacore_available:
     from ._casacore import build_table
     from ._casacore import open_table
@@ -15,7 +16,16 @@ elif _is_casacore_available:
     from ._casacore import convert_str_angle_to_rad
     from ._casacore import put_table_keyword
     from ._casacore import datestr2mjd
+    from ._casacore import mjd2datetime
 else:
     raise ModuleNotFoundError("Neither casatools or python-casacore is available")
 
-__all__ = ["build_table", "open_table", "_table", "convert_str_angle_to_rad", "put_table_keyword"]
+__all__ = [
+    "build_table",
+    "open_table",
+    "_table",
+    "convert_str_angle_to_rad",
+    "put_table_keyword",
+    "datestr2mjd",
+    "mjd2datetime"
+]
