+++ tests/ms2_filler/test_h40.py
@@ -1,6 +1,7 @@
 import datetime
 import os
 import shutil
+import time
 
 import numpy as np
 import pytest
@@ -11,20 +12,17 @@ except Exception:
     _ms = None
 
 import nro45data.psw as psw
-from nro45data.psw.ms2._casa import open_table
-
-from .. import _utils
+from nro45data.psw.ms2._casa import open_table, mjd2datetime
 
 
 @pytest.fixture(scope="module")
-def msfile():
+def msfile(data_dir):
     # input nqm file
-    data_dir = _utils._get_data_dir()
     nqmfile = "nmlh40.240926005833.01.nqm"
 
     # generate MS
-    msfile = _utils._generate_random_name()
     nqmpath = os.path.join(data_dir, nqmfile)
+    msfile = ".".join([nqmpath, time.strftime("%Y%M%dT%H%M%S"), 'ms'])
     try:
         psw.nqm2ms2(nqmpath, msfile)
     except Exception as e:
@@ -60,11 +58,11 @@ def test_ms2_structure_h40(msfile):
         # start time: 2024/09/26 00:58:54
         # end time: 2024/09/26 01:03:50
         start_expected = datetime.datetime(2024, 9, 26, 0, 58, 54)
-        start_time = _utils.mjd2datetime(time_range[0])
+        start_time = mjd2datetime(time_range[0])
         assert start_time == start_expected
 
         end_expected = datetime.datetime(2024, 9, 26, 1, 3, 50)
-        end_time = _utils.mjd2datetime(time_range[1])
+        end_time = mjd2datetime(time_range[1])
         assert end_time == end_expected
 
     with open_table(os.path.join(msfile, "SPECTRAL_WINDOW")) as tb:
@@ -109,12 +107,12 @@ def test_ms2_structure_h40(msfile):
 
         # start time: 2024/9/26 0:59:19, integration time 1sec
         start_expected = datetime.datetime(2024, 9, 26, 0, 59, 19, 500000)
-        start_time = _utils.mjd2datetime(tb.getcell("TIME", 0))
+        start_time = mjd2datetime(tb.getcell("TIME", 0))
         assert start_time == start_expected
 
         # end time: 2024/9/26 1:3:50, integration time 5sec
         end_expected = datetime.datetime(2024, 9, 26, 1, 3, 47, 500000)
-        end_time = _utils.mjd2datetime(tb.getcell("TIME", nrows - 1))
+        end_time = mjd2datetime(tb.getcell("TIME", nrows - 1))
         assert end_time == end_expected
 
         # data cell shape
